Added enemy ai

+++ Time Clash Chronicles/Assets/Scripts/Leader.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 public class Leader
 {
     public string name;
-    public int health;
+    public float health;
     public Sprite image;
 
     public Sprite flag;
 
     public Sprite border;
 
-    public Leader(string name, int health, Sprite image, Sprite flag, Sprite border)
+    public Leader(string name, float health, Sprite image, Sprite flag, Sprite border)
     {
         this.name = name;
         this.health = health;
